refactor: single deployment
