+++ pkg/decoder/object.go
@@ -1,7 +1,7 @@
 // Copyright AGNTCY Contributors (https://github.com/agntcy)
 // SPDX-License-Identifier: Apache-2.0
 
-package decoding
+package decoder
 
 import (
 	"encoding/json"
