+++ pkg/validator/schema.go
@@ -1,7 +1,7 @@
 // Copyright AGNTCY Contributors (https://github.com/agntcy)
 // SPDX-License-Identifier: Apache-2.0
 
-package validation
+package validator
 
 import (
 	"embed"
