+++ server/controller/decoding/v1/decoding.go
@@ -8,9 +8,9 @@ import (
 	"fmt"
 	"log/slog"
 
-	decodingv1grpc "buf.build/gen/go/agntcy/oasf-sdk/grpc/go/decoding/v1/decodingv1grpc"
+	"buf.build/gen/go/agntcy/oasf-sdk/grpc/go/decoding/v1/decodingv1grpc"
 	decodingv1 "buf.build/gen/go/agntcy/oasf-sdk/protocolbuffers/go/decoding/v1"
-	"github.com/agntcy/oasf-sdk/pkg/decoding"
+	"github.com/agntcy/oasf-sdk/pkg/decoder"
 )
 
 type decodingCtrl struct{}
@@ -22,7 +22,7 @@ func New() decodingv1grpc.DecodingServiceServer {
 func (t *decodingCtrl) DecodeRecord(_ context.Context, req *decodingv1.DecodeRecordRequest) (*decodingv1.DecodeRecordResponse, error) {
 	slog.Info("Received DecodeRecord request", "request", req)
 
-	res, err := decoding.DecodeRecord(req.GetRecord())
+	res, err := decoder.DecodeRecord(req.GetRecord())
 	if err != nil {
 		return nil, fmt.Errorf("failed to decode record: %w", err)
 	}
